libtypec: return values, cable prop bug fix, cmake update
Updating libtypec to run on Chrome OS. In current libtypec, lstypec
relies on return values from int functions which are used to get data
about typec devices. But, the functions which get typec device data have
no return defined which prevents the lstypec main function from
completing. This update adds return values to some functions in
libtypec.c and libtypec_sysfs_ops to resolve the issue.

Additionally, one cable prop variable is used to hold data for multiple
ports, but if a port does not have a cable, the function to get cable
prop data will immediately return without updating the cable props. This
results in cable information mistakenly being carried over from port to
port if it is not overwritten. This update includes a fix for that.

Lastly, the utils CMakeLists.txt file assumes that libtypec is in a
directory named libtypec which is not true for cloned releases. This
update removes that assumption.