libtypec: harden get_os_name, ensure it gets the right data and fix file leak
Currently get_os_name returns an offset into a buffer that is on the stack
which can then be subsequently clobbered by stack usage. Fix this by making
the buffer static, this is OK since the code is not re-entrant. The file
is not being closed, so fix this file descriptor leak by closing it. Also
ensure the buffer being parsed is terminated by a zero don't assume that
there is a trailing '\n' in the string (in case the line being read is
larger than the buffer).

One final fix is requred, when the data is of the following ordering:

VERSION_ID="11"
ID=debian

..the current implementation matches with VERSION_ID= and not ID= so
fix this by comparing strings from the start of the buffer.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>