utils: fix some array overflow issues
There are two issues that cause array index overflows, one is that
the call to print_vdo is passing an array size of 7 elements when
in fact the arrays are only 6 elements. The second issue is that
array pd3p1_fixed_supply_src_field_desc is 1 element too small.
Fix these.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>