lstypec: Add decoding for PDO voltages and other units
`lstypec` currently displays PDO fields such as "Voltage in 50mV units" as-is, requiring manual calculations to get useful values out of them.

This patch adds a decoder for such fields, listing a converted value (based on the raw value and the unit step listed in the field name) after the raw value. It's a bit rudimentary, but it works at least for the currently implemented field types.